# WSGmed/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Centralized API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/recommendations": {
            "get": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Get all recommendations",
                "description": "Get all recommendations\n\nReturns a list of all recommendations available in the system.",
                "operationId": "getRecommendations",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Check-up Visit"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Recommended check-up visit in 3 months"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-05-13T14:30:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-05-13T14:30:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Create a new recommendation",
                "description": "Create a new recommendation\n\nStores a new recommendation in the database.",
                "operationId": "storeRecommendation",
                "requestBody": {
                    "description": "Recommendation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Recommendation title",
                                        "type": "string",
                                        "example": "Check-up Visit"
                                    },
                                    "description": {
                                        "description": "Recommendation description",
                                        "type": "string",
                                        "example": "Recommended check-up visit in 3 months"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recommendation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Check-up Visit"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Recommended check-up visit in 3 months"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-13T14:30:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-13T14:30:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "title": [
                                                    "The title field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/recommendations/{id}": {
            "get": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Get recommendation details",
                "description": "Get recommendation details\n\nReturns detailed information about a specific recommendation.",
                "operationId": "getRecommendationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the recommendation",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Check-up Visit"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Recommended check-up visit in 3 months"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-13T14:30:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-13T14:30:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recommendation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recommendation not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Update a recommendation",
                "description": "Update a recommendation\n\nUpdates an existing recommendation with new data.",
                "operationId": "updateRecommendation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the recommendation",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated recommendation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Modified Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Modified recommendation description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recommendation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Modified Title"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Modified recommendation description"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-13T14:30:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-13T15:45:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recommendation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recommendation not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Delete a recommendation",
                "description": "Delete a recommendation\n\nRemoves a specific recommendation from the system.",
                "operationId": "deleteRecommendation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the recommendation",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Recommendation deleted successfully"
                    },
                    "404": {
                        "description": "Recommendation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recommendation not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your JWT token in the format: Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}